+++ EssentialFeed/Feed Presentation/FeedImagePresenter.swift
@@ -1,0 +1,67 @@
+//
+//  FeedImagePresenter.swift
+//  EssentialFeed
+//
+//  Created by Christos Tzimas on 13/7/25.
+//
+
+import Foundation
+
+public protocol FeedImageView<Image> {
+    associatedtype Image
+
+    func display(_ model: FeedImageViewModel<Image>)
+}
+
+public struct FeedImageViewModel<Image> {
+    
+    public let description: String?
+    public let location: String?
+    public let image: Image?
+    public let isLoading: Bool
+    public let shouldRetry: Bool
+    
+    public var hasLocation: Bool {
+        location != nil
+    }
+}
+
+public final class FeedImagePresenter<View: FeedImageView, Image> where View.Image == Image {
+    
+    private let view: View
+    private let imageTransformer: (Data) -> Image?
+
+    public init(view: View, imageTransformer: @escaping (Data) -> Image?) {
+        self.view = view
+        self.imageTransformer = imageTransformer
+    }
+    
+    public func didStartLoadingImageData(for model: FeedImage) {
+        view.display(FeedImageViewModel(
+            description: model.description,
+            location: model.location,
+            image: nil,
+            isLoading: true,
+            shouldRetry: false))
+    }
+    
+    public func didFinishLoadingImageData(with data: Data, for model: FeedImage) {
+        let image = imageTransformer(data)
+        let shouldRetry = image == nil
+        view.display(FeedImageViewModel(
+            description: model.description,
+            location: model.location,
+            image: image,
+            isLoading: false,
+            shouldRetry: shouldRetry))
+    }
+    
+    public func didFinishLoadingImageData(with error: Error, for model: FeedImage) {
+        view.display(FeedImageViewModel(
+            description: model.description,
+            location: model.location,
+            image: nil,
+            isLoading: false,
+            shouldRetry: true))
+    }
+}
