+++ EssentialFeedTests/Feed Presentation/FeedImagePresenterTests.swift
@@ -8,65 +8,6 @@
 import XCTest
 import EssentialFeed
 
-protocol FeedImageView<Image> {
-    associatedtype Image
-
-    func display(_ model: FeedImageViewModel<Image>)
-}
-
-struct FeedImageViewModel<Image> {
-    
-    let description: String?
-    let location: String?
-    let image: Image?
-    let isLoading: Bool
-    let shouldRetry: Bool
-    
-    var hasLocation: Bool {
-        location != nil
-    }
-}
-
-final class FeedImagePresenter<View: FeedImageView, Image> where View.Image == Image {
-    
-    private let view: View
-    private let imageTransformer: (Data) -> Image?
-
-    init(view: View, imageTransformer: @escaping (Data) -> Image?) {
-        self.view = view
-        self.imageTransformer = imageTransformer
-    }
-    
-    func didStartLoadingImageData(for model: FeedImage) {
-        view.display(FeedImageViewModel(
-            description: model.description,
-            location: model.location,
-            image: nil,
-            isLoading: true,
-            shouldRetry: false))
-    }
-    
-    func didFinishLoadingImageData(with data: Data, for model: FeedImage) {
-        let image = imageTransformer(data)
-        let shouldRetry = image == nil
-        view.display(FeedImageViewModel(
-            description: model.description,
-            location: model.location,
-            image: image,
-            isLoading: false,
-            shouldRetry: shouldRetry))
-    }
-    
-    func didFinishLoadingImageData(with error: Error, for model: FeedImage) {
-        view.display(FeedImageViewModel(
-            description: model.description,
-            location: model.location,
-            image: nil,
-            isLoading: false,
-            shouldRetry: true))
-    }
-}
-
 final class FeedImagePresenterTests: XCTestCase {
 
     func test_init_doesNotSendMessagesToView() {
